Fixed shader compilation error with low GPU

+++ src/main/java/Input/SizeListener.java
@@ -8,6 +8,6 @@ public class SizeListener {
     public static void resizeCallback(long window, int width, int height)
     {
         glViewport(0, 0, width, height);
-        Window.get().setRation((double) width /height);
+        Window.get().setRation((float) width /height);
     }
 }
